requirements.txt
tensorflow
pandas
numpy
pyarrow
scikit-learn
category_encoders
azureml-core
dotenv-python